refactor(tests): move Event repository init tests to main EventRepository test file
